+++ lib/cloudinary/api.rb
@@ -52,6 +52,15 @@ class Cloudinary::Api
     call_api(:get, uri, only(options, :next_cursor, :max_results, :tags, :context, :moderations, :direction), options)
   end
 
+  def self.resources_by_context(key, value=nil, options={})
+    resource_type = options[:resource_type] || "image"
+    uri           = "resources/#{resource_type}/context"
+    params = only(options, :next_cursor, :max_results, :tags, :context, :moderations, :direction,:key,:value)
+    params[:key] = key
+    params[:value] = value
+    call_api(:get, uri, params, options)
+  end
+
   def self.resources_by_ids(public_ids, options={})
     resource_type = options[:resource_type] || "image"
     type          = options[:type] || "upload"
@@ -98,32 +107,48 @@ class Cloudinary::Api
     resource_type = options[:resource_type] || "image"
     type          = options[:type] || "upload"
     uri           = "resources/#{resource_type}/#{type}"
-    call_api(:delete, uri, { :public_ids => public_ids }.merge(only(options, :keep_original, :invalidate)), options)
+    call_api(:delete, uri, delete_resource_params(options, :public_ids => public_ids ), options)
   end
 
   def self.delete_resources_by_prefix(prefix, options={})
     resource_type = options[:resource_type] || "image"
     type          = options[:type] || "upload"
     uri           = "resources/#{resource_type}/#{type}"
-    call_api(:delete, uri, { :prefix => prefix }.merge(only(options, :keep_original, :next_cursor, :invalidate)), options)
+    call_api(:delete, uri, delete_resource_params(options, :prefix => prefix), options)
   end
 
   def self.delete_all_resources(options={})
     resource_type = options[:resource_type] || "image"
     type          = options[:type] || "upload"
     uri           = "resources/#{resource_type}/#{type}"
-    call_api(:delete, uri, { :all => true }.merge(only(options, :keep_original, :next_cursor, :invalidate)), options)
+    call_api(:delete, uri, delete_resource_params(options, :all => true ), options)
   end
 
   def self.delete_resources_by_tag(tag, options={})
     resource_type = options[:resource_type] || "image"
     uri           = "resources/#{resource_type}/tags/#{tag}"
-    call_api(:delete, uri, only(options, :keep_original, :next_cursor, :invalidate), options)
+    call_api(:delete, uri, delete_resource_params(options), options)
   end
 
   def self.delete_derived_resources(derived_resource_ids, options={})
     uri = "derived_resources"
     call_api(:delete, uri, { :derived_resource_ids => derived_resource_ids }, options)
+  end
+
+  # Delete derived resources identified by transformation for the provided public_ids
+  # @param [String|Array] public_ids The resources the derived resources belong to
+  # @param [String|Hash|Array] transformations the transformation(s) associated with the derived resources
+  # @param [Hash] options
+  # @option options [String] :resource_type ("image")
+  # @option options [String] :type ("upload")
+  def self.delete_derived_by_transformation(public_ids, transformations, options={})
+    resource_type = options[:resource_type] || "image"
+    type          = options[:type] || "upload"
+    uri           = "resources/#{resource_type}/#{type}"
+    params = {:public_ids => public_ids}.merge(only(options, :invalidate))
+    params[:keep_original] = true
+    params[:transformations] = Cloudinary::Utils.build_eager(transformations)
+    call_api(:delete, uri, params, options)
   end
 
   def self.tags(options={})
@@ -137,7 +162,7 @@ class Cloudinary::Api
   end
 
   def self.transformation(transformation, options={})
-    call_api(:get, "transformations/#{transformation_string(transformation)}", only(options, :max_results, :next_cursor), options)
+    call_api(:get, "transformations/#{transformation_string(transformation)}", only(options, :next_cursor, :max_results), options)
   end
 
   def self.delete_transformation(transformation, options={})
@@ -213,6 +238,71 @@ class Cloudinary::Api
     call_api(:post, 'upload_mappings', params, options)
   end
 
+  def self.create_streaming_profile(name, options={})
+      params = only(options, :display_name, :representations)
+      params[:representations] = params[:representations].map do |r|
+        {:transformation => Cloudinary::Utils.generate_transformation_string(r[:transformation])}
+      end.to_json
+      params[:name] = name
+      call_api(:post, 'streaming_profiles', params, options)
+  end
+
+  def self.list_streaming_profiles
+    call_api(:get, 'streaming_profiles', {}, {})
+  end
+
+  def self.delete_streaming_profile(name, options={})
+    call_api(:delete, "streaming_profiles/#{name}", {}, options)
+  end
+
+  def self.get_streaming_profile(name, options={})
+    call_api(:get, "streaming_profiles/#{name}", {}, options)
+  end
+
+  def self.update_streaming_profile(name, options={})
+    params = only(options, :display_name, :representations)
+    params[:representations] = params[:representations].map do |r|
+      {:transformation => Cloudinary::Utils.generate_transformation_string(r[:transformation])}
+    end.to_json
+    call_api(:put, "streaming_profiles/#{name}", params, options)
+  end
+
+  # Update resources access mode. Resources are selected by the prefix
+  # @param [String] access_mode the access mode to set the resources to
+  # @param [String] prefix The prefix by which to filter applicable resources
+  # @param [Object] options    additional options
+  # @option options [String] :resource_type ("image") the type of resources to modify
+  # @option options [Fixnum] :max_results (nil) the maximum resources to process in a single invocation
+  # @option options [String] :next_cursor (nil) provided by a previous call to the method
+  def self.update_resources_access_mode_by_prefix(access_mode, prefix, options = {})
+
+      update_resources_access_mode(access_mode, :prefix, prefix, options)
+  end
+
+  # Update resources access mode. Resources are selected by the tag
+  # @param [String] access_mode the access mode to set the resources to
+  # @param [String] tag the tag by which to filter applicable resources
+  # @param [Object] options    additional options
+  # @option options [String] :resource_type ("image") the type of resources to modify
+  # @option options [Fixnum] :max_results (nil) the maximum resources to process in a single invocation
+  # @option options [String] :next_cursor (nil) provided by a previous call to the method
+  def self.update_resources_access_mode_by_tag(access_mode, tag, options = {})
+
+      update_resources_access_mode(access_mode, :tag, tag, options)
+  end
+
+  # Update resources access mode. Resources are selected by the provided public_ids
+  # @param [String] access_mode the access mode to set the resources to
+  # @param [Array<String>] public_ids The prefix by which to filter applicable resources
+  # @param [Object] options    additional options
+  # @option options [String] :resource_type ("image") the type of resources to modify
+  # @option options [Fixnum] :max_results (nil) the maximum resources to process in a single invocation
+  # @option options [String] :next_cursor (nil) provided by a previous call to the method
+  def self.update_resources_access_mode_by_ids(access_mode, public_ids, options = {})
+
+      update_resources_access_mode(access_mode, :public_ids, public_ids, options)
+  end
+
   protected
 
   def self.call_api(method, uri, params, options)
@@ -259,7 +349,21 @@ class Cloudinary::Api
     result
   end
 
+  def self.delete_resource_params(options, params ={})
+    params.merge(only(options, :keep_original, :next_cursor, :invalidate, :transformations))
+  end
+
   def self.transformation_string(transformation)
     transformation.is_a?(String) ? transformation : Cloudinary::Utils.generate_transformation_string(transformation.clone)
   end
+
+  def self.update_resources_access_mode(access_mode, by_key, value, options = {})
+    resource_type = options[:resource_type] || "image"
+    type = options[:type] || "upload"
+    params = only(options, :next_cursor)
+    params[:access_mode] = access_mode
+    params[by_key] = value
+    call_api("post", "resources/#{resource_type}/#{type}/update_access_mode", params, options)
+  end
+
 end