+++ samples/photo_album/app/models/photo.rb
@@ -1,6 +1,9 @@
 class Photo < ActiveRecord::Base
   attr_accessible :title, :bytes, :image, :image_cache
+
+  belongs_to :album
+
   mount_uploader :image, ImageUploader
-  
+
   validates_presence_of :title, :image
 end