+++ spec/spec_helper.rb
@@ -4,6 +4,10 @@ require 'rspec/version'
 require 'rest_client'
 
 TEST_IMAGE_URL = "http://cloudinary.com/images/old_logo.png"
+TEST_IMG = "spec/logo.png"
+TEST_IMG_W = 241
+TEST_IMG_H = 51
+
 TEST_TAG = 'cloudinary_gem_test'
 TIMESTAMP_TAG = "#{TEST_TAG}_#{rand(999999999)}_#{RUBY_VERSION}_#{ defined? Rails::version ? Rails::version : 'no_rails'}"
 
@@ -166,7 +170,7 @@ end
 RSpec::Matchers.define :deep_hash_value do |expected|
   match do |actual|
     expected.all? do |path, value|
-      Cloudinary.values_match? deep_fetch(actual, path), value
+      Cloudinary.values_match? value, deep_fetch(actual, path)
     end
   end
 end
@@ -175,7 +179,7 @@ RSpec::Matchers.alias_matcher :have_deep_hash_values_of, :deep_hash_value
 
 module Cloudinary
   # @api private
-  def self.values_match?( actual, expected)
+  def self.values_match?(expected, actual)
     if Hash === actual
       return hashes_match?(expected, actual) if Hash === expected
     elsif Array === expected && Enumerable === actual && !(Struct === actual)